Refactor boolean flag out of perform_validation
I moved the regex up to be a constant in part to shorten the line but
also because I think its purpose in the logic is a little clearer as a
named constant.

Mostly, I feel like `had_error = false` and `valid_so_far = true` et al.
are a smell, and since one of the tests already raises an exception, I
felt like it was easy enough to deal with. If it bothers us for a test
to raise an immediately-caught ArgumentError, I guess we could create
our own exception class and catch that, or something else, but that felt
like overkill to me.

Replacing `result = []; ... return result` with `tap` is pretty much a
stylistic decision I made while I was here.